requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
tiktoken==0.5.2

# Web UI (optional)
streamlit==1.29.0
